requirements.txt
google-genai==1.43.0